+++ packages/core/useLazyScript/demo.vue
@@ -1,8 +1,19 @@
 <script lang="ts" setup>
 import { useLazyScript } from './index'
-const { load, unload } = useLazyScript('https://unpkg.com/vue@3/dist/vue.global.js')
-load.then(()=>{
+const load = useLazyScript('https://unpkg.com/vue@3/dist/vue.global.js')
+load.then(() => {
+    console.log('js加载完成')
+})
+
+const loads = useLazyScript([
+    'https://unpkg.com/vue@3/dist/vue.global.js',
+    'https://unpkg.com/vue-router@4'
+])
+
+loads.then(() => {
     console.log('js加载完成')
 })
 </script>
-<template></template>
+<template>
+    <div>1212</div>
+</template>