+++ packages/core/useThrottleFn/index.ts
@@ -1,29 +1,45 @@
-function createWrapper<T extends (...args: any[]) => any>(executor: (...args: any[]) => any, fn: T) {
-    return function () {
-        return new Promise<Awaited<ReturnType<T>>>((resolve, reject) => {
-            Promise.resolve(executor(fn)).then(resolve, reject)
-        })
-    }
-}
 function throttleFn<T extends (...args: any[]) => any>(ms?: number) {
-    let lastExe = 0
-    let isLock = false
-    let lastValue: any
-    return function (fn: T) {
-        const elapsed = Date.now() - lastExe
-        if (typeof ms === 'number' && elapsed > ms) {
-            lastExe = Date.now()
-            lastValue = fn()
-        }
-        if (typeof ms === 'undefined' && !isLock) {
-            lastValue = fn(() => {
-                isLock = false
-            })
-            isLock = true
-        }
-        return lastValue
+  let lastExe = 0;
+  let isLock = false;
+  let lastValue: any;
+  let timer: ReturnType<typeof setTimeout> | undefined;
+  const clear = () => {
+    if (timer) {
+      clearTimeout(timer);
+      timer = undefined;
     }
+  };
+
+  return function (fn: T, trailing: boolean, leading: boolean) {
+    const elapsed = Date.now() - lastExe;
+    clear();
+    if (typeof ms === "number" && elapsed >= ms && leading) {
+      lastExe = Date.now();
+      lastValue = fn();
+    } else if (typeof ms === "number" && trailing && elapsed < ms) {
+      lastValue = new Promise((resolve, reject) => {
+        timer = setTimeout(() => {
+          lastExe = Date.now();
+          resolve(fn());
+          clear();
+        }, ms - elapsed);
+      });
+    }
+    if (typeof ms === "undefined" && !isLock) {
+      lastValue = fn(() => {
+        isLock = false;
+      });
+      isLock = true;
+    }
+    return lastValue;
+  };
 }
-export function useThrottleFn<T extends (...args: any[]) => any>(fn: T, ms?: number) {
-    return createWrapper<T>(throttleFn<T>(ms), fn)
+
+export function useThrottleFn<T extends (...args: any[]) => any>(
+  fn: T,
+  ms?: number,
+  trailing = true,
+  leading = true
+) {
+  return throttleFn<T>(ms)(fn, trailing, leading);
 }